[Chore] Add blueprints for better organisation
