+++ app/v1/auth/__init__.py
@@ -1,0 +1,8 @@
+# v1/auth/__init__.py
+
+
+from flask import Blueprint
+
+auth = Blueprint('auth', __name__)
+
+from . import views
