+++ app/v1/menu/__init__.py
@@ -1,0 +1,7 @@
+# menu/__init__.py
+
+from flask import Blueprint
+
+menu = Blueprint('menu',__name__)
+
+from . import views
